+++ Modules/Biometric/Database/Migrations/2025_06_03_113406_add_clockin_method_biometric_employees_table.php
@@ -12,14 +12,13 @@ return new class extends Migration
      */
     public function up(): void
     {
-        if (! Schema::hasTable('biometric_employees')) {
-            return;
-        }
-        if (!Schema::hasColumn('biometric_employees', 'force_biometric_clockin')) {
-            Schema::table('biometric_employees', function (Blueprint $table) {
+        if (Schema::hasTable('biometric_employees')) {
+        Schema::table('biometric_employees', function (Blueprint $table) {
+            if (!Schema::hasColumn('biometric_employees', 'force_biometric_clockin')) {
                 $table->boolean('force_biometric_clockin')->default(true)->after('company_id');
-            });
-        }
+            }
+        });
+    }
     }
 
     /**
@@ -27,8 +26,12 @@ return new class extends Migration
      */
     public function down(): void
     {
+        if (Schema::hasTable('biometric_employees')) {
         Schema::table('biometric_employees', function (Blueprint $table) {
-            $table->dropColumn('force_biometric_clockin');
+            if (Schema::hasColumn('biometric_employees', 'force_biometric_clockin')) {
+                $table->dropColumn('force_biometric_clockin');
+            }
         });
+    }
     }
 };
