+++ Modules/BusinessSettingsModule/Database/Migrations/2024_06_02_173558_add_to_col_package_subscriber_table.php
@@ -13,14 +13,13 @@ class AddToColPackageSubscriberTable extends Migration
      */
     public function up()
     {
-        if (! Schema::hasTable('package_subscribers')) {
-            return;
-        }
-        if (!Schema::hasColumn('package_subscribers', 'is_canceled')) {
-            Schema::table('package_subscribers', function (Blueprint $table) {
+        if (Schema::hasTable('package_subscribers')) {
+        Schema::table('package_subscribers', function (Blueprint $table) {
+            if (!Schema::hasColumn('package_subscribers', 'is_canceled')) {
                 $table->tinyInteger('is_canceled')->default(0);
-            });
-        }
+            }
+        });
+    }
     }
 
     /**
@@ -30,8 +29,12 @@ class AddToColPackageSubscriberTable extends Migration
      */
     public function down()
     {
+        if (Schema::hasTable('package_subscribers')) {
         Schema::table('package_subscribers', function (Blueprint $table) {
-            $table->dropColumn('is_canceled');
+            if (Schema::hasColumn('package_subscribers', 'is_canceled')) {
+                $table->dropColumn('is_canceled');
+            }
         });
+    }
     }
 }
