+++ Modules/BusinessSettingsModule/Database/Migrations/2024_06_06_160149_add_to_col_transaction_id_to_package_subscriber_logs_tabel.php
@@ -13,13 +13,8 @@ class AddToColTransactionIdToPackageSubscriberLogsTabel extends Migration
      */
     public function up()
     {
-        if (! Schema::hasTable('package_subscriber_logs')) {
-            return;
-        }
         Schema::table('package_subscriber_logs', function (Blueprint $table) {
-            if (! Schema::hasColumn('package_subscriber_logs', 'primary_transaction_id')) {
-                $table->foreignUuid('primary_transaction_id')->default(0)->nullable();
-            }
+            $table->foreignUuid('primary_transaction_id')->default(0);
         });
     }
 
@@ -30,8 +25,12 @@ class AddToColTransactionIdToPackageSubscriberLogsTabel extends Migration
      */
     public function down()
     {
+        if (Schema::hasTable('package_subscriber_logs')) {
         Schema::table('package_subscriber_logs', function (Blueprint $table) {
-            $table->dropColumn('primary_transaction_id');
+            if (Schema::hasColumn('package_subscriber_logs', 'primary_transaction_id')) {
+                $table->dropColumn('primary_transaction_id');
+            }
         });
+    }
     }
 }
