+++ Modules/BusinessSettingsModule/Database/Migrations/2024_09_25_133132_add_to_col_cron_jobs_table.php
@@ -13,14 +13,13 @@ class AddToColCronJobsTable extends Migration
      */
     public function up()
     {
-        if (! Schema::hasTable('cron_jobs')) {
-            return;
-        }
-        if (!Schema::hasColumn('cron_jobs', 'type')) {
-            Schema::table('cron_jobs', function (Blueprint $table) {
+        if (Schema::hasTable('cron_jobs')) {
+        Schema::table('cron_jobs', function (Blueprint $table) {
+            if (!Schema::hasColumn('cron_jobs', 'type')) {
                 $table->string('type')->nullable();
-            });
-        }
+            }
+        });
+    }
     }
 
     /**
@@ -30,8 +29,12 @@ class AddToColCronJobsTable extends Migration
      */
     public function down()
     {
+        if (Schema::hasTable('cron_jobs')) {
         Schema::table('cron_jobs', function (Blueprint $table) {
-            $table->dropColumn('type');
+            if (Schema::hasColumn('cron_jobs', 'type')) {
+                $table->dropColumn('type');
+            }
         });
+    }
     }
 }
