+++ Modules/Performance/Database/Migrations/2025_02_18_061400_add_columns_in_performance_settings_table.php
@@ -13,16 +13,19 @@ return new class extends Migration
      */
     public function up(): void
     {
-        if (! Schema::hasTable('performance_settings')) {
-            return;
-        }
-        if (!Schema::hasColumn('performance_settings', 'send_slack_notification')) {
-            Schema::table('performance_settings', function (Blueprint $table) {
+        if (Schema::hasTable('performance_settings')) {
+        Schema::table('performance_settings', function (Blueprint $table) {
+            if (!Schema::hasColumn('performance_settings', 'send_slack_notification')) {
                 $table->enum('send_slack_notification', ['yes', 'no'])->default('no')->after('view_meeting_participant');
+            }
+            if (!Schema::hasColumn('performance_settings', 'send_push_notification')) {
                 $table->enum('send_push_notification', ['yes', 'no'])->default('no')->after('send_slack_notification');
+            }
+            if (!Schema::hasColumn('performance_settings', 'send_email_notification')) {
                 $table->enum('send_email_notification', ['yes', 'no'])->default('no')->after('send_push_notification');
-            });
-        }
+            }
+        });
+    }
 
         $performanceSettings = PerformanceSetting::all();
 
@@ -33,9 +36,13 @@ return new class extends Migration
             }
         }
 
+        if (Schema::hasTable('performance_settings')) {
         Schema::table('performance_settings', function (Blueprint $table) {
-            $table->dropColumn('send_notification');
+            if (Schema::hasColumn('performance_settings', 'send_notification')) {
+                $table->dropColumn('send_notification');
+            }
         });
+    }
     }
 
     /**
@@ -43,11 +50,19 @@ return new class extends Migration
      */
     public function down(): void
     {
+        if (Schema::hasTable('performance_settings')) {
         Schema::table('performance_settings', function (Blueprint $table) {
-            $table->dropColumn('send_slack_notification');
-            $table->dropColumn('send_push_notification');
-            $table->dropColumn('send_email_notification');
+            if (Schema::hasColumn('performance_settings', 'send_slack_notification')) {
+                $table->dropColumn('send_slack_notification');
+            }
+            if (Schema::hasColumn('performance_settings', 'send_push_notification')) {
+                $table->dropColumn('send_push_notification');
+            }
+            if (Schema::hasColumn('performance_settings', 'send_email_notification')) {
+                $table->dropColumn('send_email_notification');
+            }
         });
+    }
 
         $performanceSettings = PerformanceSetting::all();
 
@@ -58,11 +73,13 @@ return new class extends Migration
             }
         }
 
+        if (Schema::hasTable('performance_settings')) {
         Schema::table('performance_settings', function (Blueprint $table) {
-            if (! Schema::hasColumn('performance_settings', 'send_notification')) {
-                $table->enum('send_notification', ['yes', 'no'])->default('no')->after('company_id')->nullable();
+            if (!Schema::hasColumn('performance_settings', 'send_notification')) {
+                $table->enum('send_notification', ['yes', 'no'])->default('no')->after('company_id');
             }
         });
+    }
     }
 
 };
