+++ Modules/ProviderManagement/Database/Migrations/2022_06_22_090257_column_add_to_withdraw_request_table.php
@@ -13,15 +13,16 @@ class ColumnAddToWithdrawRequestTable extends Migration
      */
     public function up()
     {
-        if (! Schema::hasTable('withdraw_requests')) {
-            return;
-        }
-        if (!Schema::hasColumn('withdraw_requests', 'is_paid')) {
-            Schema::table('withdraw_requests', function (Blueprint $table) {
+        if (Schema::hasTable('withdraw_requests')) {
+        Schema::table('withdraw_requests', function (Blueprint $table) {
+            if (!Schema::hasColumn('withdraw_requests', 'is_paid')) {
                 $table->boolean('is_paid')->default(0);
+            }
+            if (!Schema::hasColumn('withdraw_requests', 'note')) {
                 $table->string('note')->nullable();
-            });
-        }
+            }
+        });
+    }
     }
 
     /**
@@ -31,9 +32,15 @@ class ColumnAddToWithdrawRequestTable extends Migration
      */
     public function down()
     {
+        if (Schema::hasTable('withdraw_requests')) {
         Schema::table('withdraw_requests', function (Blueprint $table) {
-            $table->dropColumn('is_paid');
-            $table->dropColumn('note');
+            if (Schema::hasColumn('withdraw_requests', 'is_paid')) {
+                $table->dropColumn('is_paid');
+            }
+            if (Schema::hasColumn('withdraw_requests', 'note')) {
+                $table->dropColumn('note');
+            }
         });
+    }
     }
 }
