+++ Modules/ProviderManagement/Database/Migrations/2022_09_21_235326_col_add_to_withdraw_requests_table.php
@@ -13,14 +13,13 @@ class ColAddToWithdrawRequestsTable extends Migration
      */
     public function up()
     {
-        if (! Schema::hasTable('withdraw_requests')) {
-            return;
-        }
-        if (!Schema::hasColumn('withdraw_requests', 'admin_note')) {
-            Schema::table('withdraw_requests', function (Blueprint $table) {
+        if (Schema::hasTable('withdraw_requests')) {
+        Schema::table('withdraw_requests', function (Blueprint $table) {
+            if (!Schema::hasColumn('withdraw_requests', 'admin_note')) {
                 $table->string('admin_note')->nullable();
-            });
-        }
+            }
+        });
+    }
     }
 
     /**
@@ -30,8 +29,12 @@ class ColAddToWithdrawRequestsTable extends Migration
      */
     public function down()
     {
+        if (Schema::hasTable('withdraw_requests')) {
         Schema::table('withdraw_requests', function (Blueprint $table) {
-            $table->dropColumn('admin_note');
+            if (Schema::hasColumn('withdraw_requests', 'admin_note')) {
+                $table->dropColumn('admin_note');
+            }
         });
+    }
     }
 }
