+++ Modules/ProviderManagement/Database/Migrations/2023_05_16_130004_add_col_to_providers_table.php
@@ -13,14 +13,13 @@ class AddColToProvidersTable extends Migration
      */
     public function up()
     {
-        if (! Schema::hasTable('providers')) {
-            return;
-        }
-        if (!Schema::hasColumn('providers', 'coordinates')) {
-            Schema::table('providers', function (Blueprint $table) {
+        if (Schema::hasTable('providers')) {
+        Schema::table('providers', function (Blueprint $table) {
+            if (!Schema::hasColumn('providers', 'coordinates')) {
                 $table->json('coordinates')->nullable();
-            });
-        }
+            }
+        });
+    }
     }
 
     /**
@@ -30,8 +29,12 @@ class AddColToProvidersTable extends Migration
      */
     public function down()
     {
+        if (Schema::hasTable('providers')) {
         Schema::table('providers', function (Blueprint $table) {
-            $table->dropColumn('coordinates');
+            if (Schema::hasColumn('providers', 'coordinates')) {
+                $table->dropColumn('coordinates');
+            }
         });
+    }
     }
 }
