+++ Modules/ProviderManagement/Database/Migrations/2023_11_19_112426_add_is_suspended_col_to_providers_table.php
@@ -13,14 +13,13 @@ class AddIsSuspendedColToProvidersTable extends Migration
      */
     public function up()
     {
-        if (! Schema::hasTable('providers')) {
-            return;
-        }
-        if (!Schema::hasColumn('providers', 'is_suspended')) {
-            Schema::table('providers', function (Blueprint $table) {
+        if (Schema::hasTable('providers')) {
+        Schema::table('providers', function (Blueprint $table) {
+            if (!Schema::hasColumn('providers', 'is_suspended')) {
                 $table->boolean('is_suspended')->default(0);
-            });
-        }
+            }
+        });
+    }
     }
 
     /**
@@ -30,8 +29,12 @@ class AddIsSuspendedColToProvidersTable extends Migration
      */
     public function down()
     {
+        if (Schema::hasTable('providers')) {
         Schema::table('providers', function (Blueprint $table) {
-            $table->dropColumn('is_suspended');
+            if (Schema::hasColumn('providers', 'is_suspended')) {
+                $table->dropColumn('is_suspended');
+            }
         });
+    }
     }
 }
