+++ Modules/Purchase/Database/Migrations/2025_06_23_124939_add_category_id_in_vendor_table.php
@@ -13,15 +13,14 @@ return new class extends Migration
     {
 
         // purchase_vendors in this table add category_id
-        if (! Schema::hasTable('purchase_vendors')) {
-            return;
-        }
-        if (!Schema::hasColumn('purchase_vendors', 'category_id')) {
-            Schema::table('purchase_vendors', function (Blueprint $table) {
+        if (Schema::hasTable('purchase_vendors')) {
+        Schema::table('purchase_vendors', function (Blueprint $table) {
+            if (!Schema::hasColumn('purchase_vendors', 'category_id')) {
                 $table->integer('category_id')->unsigned()->nullable();
-                $table->foreign('category_id')->references('id')->on('purchase_vendor_categories')->onDelete('set null');
-            });
-        }
+            }
+            $table->foreign('category_id')->references('id')->on('purchase_vendor_categories')->onDelete('set null');
+        });
+    }
     }
 
     /**
@@ -31,9 +30,13 @@ return new class extends Migration
     {
 
         // drop category_id from purchase_vendors table
+        if (Schema::hasTable('purchase_vendors')) {
         Schema::table('purchase_vendors', function (Blueprint $table) {
             $table->dropForeign(['category_id']);
-            $table->dropColumn('category_id');
+            if (Schema::hasColumn('purchase_vendors', 'category_id')) {
+                $table->dropColumn('category_id');
+            }
         });
+    }
     }
 };
