+++ Modules/ServiceManagement/Database/Migrations/2023_05_08_161525_add_col_to_services_table.php
@@ -13,14 +13,13 @@ class AddColToServicesTable extends Migration
      */
     public function up()
     {
-        if (! Schema::hasTable('services')) {
-            return;
-        }
-        if (!Schema::hasColumn('services', 'min_bidding_price')) {
-            Schema::table('services', function (Blueprint $table) {
+        if (Schema::hasTable('services')) {
+        Schema::table('services', function (Blueprint $table) {
+            if (!Schema::hasColumn('services', 'min_bidding_price')) {
                 $table->decimal('min_bidding_price', 24, 3)->default(0)->after('avg_rating');
-            });
-        }
+            }
+        });
+    }
     }
 
     /**
@@ -30,8 +29,12 @@ class AddColToServicesTable extends Migration
      */
     public function down()
     {
+        if (Schema::hasTable('services')) {
         Schema::table('services', function (Blueprint $table) {
-            $table->dropColumn('min_bidding_price');
+            if (Schema::hasColumn('services', 'min_bidding_price')) {
+                $table->dropColumn('min_bidding_price');
+            }
         });
+    }
     }
 }
